[9주차/앨빈] 워크북 제출합니다.

+++ src/main/java/javalab/umc7th_mission/domian/review/Review.java
@@ -35,7 +35,7 @@ public class Review extends BaseEntity {
 
     private String title;
 
-    private Float score;
+    private Double score;
 
     private String body;
